Replace Xbehave with BDDfy in the API component tests project
XBehave was used in the API project and it has been deprecated.  In this commit it is replaced with BDDfy.

BDDfy was chosen because, whilst it has not been updated in some time, it has not been deprecated. It is also used in the Functional Tests project, so it was chosen for consistentency.